Renamed folders for components and created index file for Button component
