+++ src/components/header/Header.jsx
@@ -3,7 +3,7 @@ import { NavLink } from "react-router-dom";
 
 import "./header.scss";
 import logo from "../../assets/images/logo.png";
-import Button from "../button/Button";
+import Button from "../Button";
 
 const Header = () => {
   return (
