+++ src/pages/modal/Modal.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Button from "../../components/button/Button";
+import Button from "../../components/Button";
 
 import "./Modal.scss";
